Modify blog deletion for authorized users who created it and return suitable status code

+++ part4/blog-list/controllers/blogs.js
@@ -18,7 +18,6 @@ blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
   const token = request.token
-  console.log(token)
 
   try {
     const decodedToken = jwt.verify(token, process.env.SECRET)
@@ -73,8 +72,23 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      response.status(404).end()
+    }
+
+    if (user._id.toString() === blog.user.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      //remove blogid from user
+      user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+      await User.findByIdAndUpdate(decodedToken.id, user)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
   } catch(exception) {
     next(exception)
   }
